Inicio da modelação de serviços OK

+++ src/oficinamecanica/model/Servico.java
@@ -35,7 +35,7 @@ public class Servico {
         return preco;
     }
 
-    public void setPreco(float preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
